Improvements for throughput and error handling.
Bump connection pool timeout to 100 (was 20), as we can hit this when we run into the 'max hosts per connection' limit. Having a big value here shouldn't hurt the typical case.

Add a cookie store to the local context when executing a GET request, as this prevents a blocking situation where HttpClient is synchronizing on a single cookie store.

Added notes about migration to HttpClient 4.1 (new APIs)