Added new "skip blocked URLs" setting
This controls whether the FetchBuffer code will block if it gets a group
of URLs to fetch, and that server is still active. If non-blocking, then
the group of URLs is not fetched, and output with a SKIPPED status.

Also added more tracing to FetchBuffer, and cleaned up the code.