Improve termination of fetch
- Add new AbortedFetchReason = INTERRUPTED
- Try harder to avoid losing URLs when doing a hard termination
- Avoid trying to use invalid collector when it's too late.
- Added test to exhibit the problem, validate the fix.

But it's still not working, because interrupted the FetchTask doesn't seem to cause HttpClient
to actually interrupt the current thread. Probably because with blocking I/O, an interrupt has
no impact. There's some confusing suggestions on the HttpClient mailing list re how to handle
this situation gracefully, but I'll need to post some more questions.